refactor : 로직 이동
